requirements.txt
fastapi
uvicorn
pydantic
requests
opencv-python-headless
adafruit-circuitpython-dht
RPi.GPIO
python-dotenv
openai
numpy
